backend/recsa_challenge/appointments/api/serializers/appoiment_serializer.py
from rest_framework import serializers
from recsa_challenge.appointments.models import Appointment

class AppointmentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Appointment
        fields = '__all__'
        read_only_fields = ('id', 'created_at', 'updated_at')